Move firmware to use RESTful principles, and create `ui/bff/firmware`.
