+++ goosebit/schema/firmware.py
