+++ goosebit/ui/bff/firmware/__init__.py
@@ -1,0 +1,1 @@
+from .routes import router  # noqa: F401
